feat(spot): complete WebSocket user data sessions
Add session.subscriptions and Ed25519 session-based user data subscriptions with reconnect reauthentication and logout-aware lifecycle handling. Keep signed subscriptions independent and align subscription identifiers to int64.

+++ Binance.Api/Spot/Socket/BinanceSpotUserDataStreamEvent.cs
@@ -9,5 +9,5 @@ public record BinanceSpotUserDataStreamEvent : BinanceSocketStreamEvent
     /// Server-assigned identifier of the user data stream subscription.
     /// </summary>
     [JsonIgnore]
-    public int SubscriptionId { get; internal set; }
+    public long SubscriptionId { get; internal set; }
 }
